Fact-checker/requirements.txt
python-telegram-bot==20.7
SQLAlchemy==2.0.25
alembic==1.13.1
fastapi==0.109.0
uvicorn==0.27.0
redis==5.0.1
celery==5.3.6
pillow==10.2.0
python-dotenv==1.0.0
transformers==4.37.1
torch==2.1.2
numpy==1.26.3
pandas==2.1.4
scikit-learn==1.4.0
requests==2.31.0
aiohttp==3.9.1
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
black==24.1.0
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
python-multipart==0.0.6
pydantic==2.5.3
opencv-python==4.9.0.80
pytesseract==0.3.10
nltk==3.8.1
gensim==4.3.2
spacy==3.7.2
